feat:added wastebank profile endpoints

+++ database/migrations/000016_create_customer_request_items.up.sql
@@ -4,6 +4,6 @@ CREATE TABLE IF NOT EXISTS customer_request_items (
     request_id UUID NOT NULL REFERENCES customer_requests(id) ON DELETE CASCADE,
     waste_type_id UUID NOT NULL REFERENCES waste_types(id) ON DELETE CASCADE,
     quantity BIGINT,
-    verified_weight FLOAT,
+    verified_weight DECIMAL,
     verified_subtotal BIGINT
 );
