+++ internal/entity/waste_bank_profile_entity.go
@@ -9,7 +9,7 @@ import (
 type WasteBankProfile struct {
 	ID               uuid.UUID `gorm:"type:uuid;default:gen_random_uuid();primaryKey"`
 	UserID           uuid.UUID `gorm:"column:user_id;unique;not null"`
-	TotalWasteWeight int64     `gorm:"column:total_waste_weight;default:0"`
+	TotalWasteWeight float64   `gorm:"column:total_waste_weight;default:0"`
 	TotalWorkers     int64     `gorm:"column:total_workers;default:0"`
 	OpenTime         time.Time `gorm:"column:open_time;type:time"`
 	CloseTime        time.Time `gorm:"column:close_time;type:time"`
