+++ internal/entity/waste_collector_profile_entity.go
@@ -8,6 +8,6 @@ type WasteCollectorProfile struct {
 	ID               uuid.UUID `gorm:"type:uuid;default:gen_random_uuid();primaryKey"`
 	UserID           uuid.UUID `gorm:"column:user_id;unique;not null"`
 	User             User      `gorm:"foreignKey:UserID"`
-	TotalWasteWeight int64     `gorm:"column:total_waste_weight;default:0"`
+	TotalWasteWeight float64   `gorm:"column:total_waste_weight;default:0"`
 	Notes            string    `gorm:"column:notes"`
 }
